Add ids into the data packages and add glue table ETL
Pretty simple update. Added a UUID into the data package generation. The 
UUID is generated in the coordinating function that generates all of the 
packages and is used in each package. So all packages generated at the 
same time should have the same id.

The glue tables ETL was also aded into the main CSV generation process 
and into the standard settings file.