Improve script logging, fix script packaging errors
Added the `coloredlogs` module to produce colorful logging output (and
easily configured loggers) to the console scripts (any module named
`cli.py` in the overall package).

Fixed an error that caused the console scripts not to be deployed on the
Travis system, breaking the builds.

Also confirmed that the instantiated database passes the current data
validation tests when run against a database generated with the current
repository (my local DB had gotten out of sync).

Progress toward #327.