Set up bandit, use SHA-256 instead of SHA-1
Set up a pre-commit hook for `bandit` which does security auditing of
Python code, and as a result replaced the no longer recommended SHA-1
hashing algorithm in with SHA-256. Also told bandit (which is used by
Codacy as well) to ignore our sadly required use of FTP to get the FERC
Form 1 and EPA CEMS data.