Finish conversion of scripts to entrypoints.
Finished converting the deployable scripts into entry points, creating a
new `convert` subpackage containing epacems_to_parquet and
ferc1_to_sqlite as modules with `main()` functions.

Also simplified the testing arrangement, reading environment variables
from the system to figure out if we're running on Travis or not, and
adjusting the scope of the data to be run through ETL accordingly.

This means we no longer need the specialized `travis_ci` test or Travis
specific database connections, and can adjust many of the tests to do
the right thing, depending on whether we're testing locally or on
Travis (or some other CI platform).

Closes #327