Remove conditional xfail from epacems_to_parquet
The `pudl.convert.epacems_to_parquet` test was failing because it relies
on the PUDL DB being populated, and one of the other tests (which checks
to see if it's possible to ingest just ferc1 without any EIA or other
data) was re-populating the PUDL DB and wiping out the data that this
test requires.  As a temporary (lol famous last words) fix, I just
changed the order of the tests so that this one happens before the
database gets clobbered by the lonely ferc1 test. Really they should be
referring to two different databases... which will be easy to arrange if
we are using file-based data storage (data packages and/or SQLite
databases).

This commit just removes the conditional xfail that I'd put around the
test to let it be tried, but not blow up the tests if it failed, on
Travis. And also adds some cleanup after the other lonely ferc1 test
(dropping all the tables from the testing database as we do with the
other tests that run `pudl.init.init_db()`)

Closes #351