Preliminary CEMS data packaging.
I was trying to do two main things here: convert the cems etl functions 
to output data packages and allow for partitioned data tables. We wanted 
to allow for CEMs to exported as CSVs broken down by year or states, so 
I enabled partitioning in general in metadata generation. I also enabled 
compression for CEMS, but the dp validation is failing, so I commented 
out two lines in `load.py` for the meantime. Hopefully we can make this  
work on the dp validation side but in the meantime we can generate cems 
data packages. This is all still pretty preliminary!! There are several 
things I have left to solidify. Also, this breaks the pudl database! For 
now, I'm not going to continue to ensure the database is fully operable 
on this branch.

Working on Issue #340.