Added linting environments to tox & other settings
the Tox setup now includes several "linters" that can check whether the
code is valid and up to standards before actually going to the trouble
of running the tests. These linters include:
  * `pre-commit`: common python mistakes outside of PEP8, yaml file syntax
  * `doc8`: all the .rst files under doc/ for validity/formatting.
  * `bandit`: for known security vulnerabilities.
  * `flake8`: code & docstring syntax & style (generally against PEP8).

To run any one of the above you can do `tox -e <name>` To run them all
together, you can run `tox -e linters`  Note that currently `flake8`
still has a big pile of errors that need to be cleaned up so running
`flake8` or `linters` will fail.

These checks are a subset of the things we're integrating into the
standard set of git pre-commit hooks (which are run only against
whatever files you're trying to commit, rather than the entire
repository as with the above linters in Tox).

Closes #362