Remove ferc1_test DB logic and clean up docstrings.
Removed references to the ferc1_test database, since we can now put
these database files wherever we want, and simply differentiate based on
location -- if it's in a /tmp folder generated by pytest or some other
disposable place it's a test database.

The --live_pudl_db pytest argument persists though, since we need to be
able to tell the system whether it should be looking at an existing
database or creating a new one.

Also fixed a bunch of docstring formatting errors that flake8 was
bringing up... trying to get the backlog of linter errors down to zero a
little bit at a time.

Closes #342 (FERC 1 testing DB logic)
Progress toward #316 (writing docstrings)