embed the pudl_engine and table metadata within PudlTabl
Now all of the output/PudlTabl functions include pudl_engine and pt as 
arguments and they now incorporated into the PudlTabl object. 
pudl_engine is a new argument to PudlTabl and the pt is generated within 
PudlTabl and are passed around to all of the incorporated functions. 
Working on Issue #356.