Inserting glue into subpackages and etl parameters into source metadata
Closes Issue #354... sort of. Instead of integrating external foreign 
keys, we integrated the glue tables into the related sub-data packages.

In order to ensure consistency when flattening the data packages, I was 
working on Issue #381. Now when a bundle of datapackages are being 
flattened for export to sqlite or otherwise, a check will be made to see 
if the etl parameters are consistent. If they aren't, the flattening 
will fail.