Add data validation infrastructure module and tests
Created a new data validation module `pudl.validate` that contains some
functions that are useful for checking whether or not the data we've
compiled looks right, and catching accidental alterations that result
from re-factoring or other changes to the code.

Started experimenting with using it in the data validation tests, and an
accompanying notebook, both of which pertain to EIA 923 Fuel Recepts and
Costs.

For now the specification of the test cases is being stored in some
dictionaries within the `pudl.validate` module, so that they can be
accessed both by the notebook, and by PyTest.

Progress toward #275