Integrating the new specs for group/partitioning
Removing the previous implemention for specs on partitioning. Making the 
metadata include "groups" for partitioned tables. Working on #364. This 
still does not work generating multi-part CEMS file because the index is 
the primary key... and with mulitple indexes starting at 0 there are non 
uniquie pk's...

Also, I removed all references to a default datapackage bundle name 
Closes #367.