Add data validation tests and notebook for frc_eia923
Created a bunch of new data validation tests for the EIA Fuel Receipts
and Costs data.

Also broke out the individual tables for EIA 923 into their own
data validation (pytest) modules and notebooks, of which this is the
first.

Note that there are two tests which currently xfail:
* coal mercury content: because of an anomalously high mercury
  population (9.0ppm), and also a majority of the data reporting 0.0ppm.
  (See issue #390)
* natural gas heat content per unit of fuel: because of a small
  population of records with heat contents 1/10th what they should be.
  (See issue #391)

Progress toward #275