Add data validation tests / notebook for bf_eia923
One new outstanding xfail-ing data validation, in which the
`fuel_type_code==BLQ` ("black liquor") is lumped in with oil in the
boiler_fuel_eia923 table (and more generally, via a list in
`constants.py` See issue #393 -- unclear why moving it from the oil
category to the waste category for the purposes of generating our
simplified `fuel_type_code_pudl` didn't work but... it didn't.

In addition to those changes, I reduced the allowable complexity from 22
to 10 (which seems "generally recommended") in the Flake8 configuration,
and also added # noqa: C901 to our existing overly complex functions...
so at least we'll be held accountable on the creation of new functions.