Mucking with autoincremental id's
I've removed all of the columns that were using the `id` column as an 
autoincremental id (with the exception of the fuel_receipts_costs_eia923 
table). Now in the pudl/package_data/meta/datapackage/datapackage.json 
file there is an additional piece of metadata called 'autoincrement', 
which contains all of the tables that need an autoincremented id when 
being loaded into sqlite.

On the fuel_receipts_costs_eia923 table, I've kept our previous use of 
the index for the id because without it there are non-unique records and 
it breaks the goodtables.validate without it.

Working on Issue #364.... nearly finished!