Update data_years for EIA923 2018 Early Release data
The EIA923 early release data is available for 2018, but it doesn't
conform to the same naming convention that we expect. However, the 2017
data has been moved into the "archive" directory... so we need the 2018
data to look like the most recent data (or we need to carve out a new
special case, ugh).

For now the data_years for eia923 includes 2018... even though it's not
actually downloadable b/c of the filename stupidity. Come *on* EIA!
Thank of the machines!