Pass arguments through to capacity_factor()
We've been truncating the distribution of capacity factors to exclude
negative and excessively high (erroneous) values, by default, but for
data validation purposes we actually need the whole thing. And it turned
out that we couldn't override the defaults, because they weren't getting
passed through from the PudlTabl class to the capacity_factor()
function. This is now fixed.