Remove sqlalchemy-postgres-copy dependency
Rather than depending on the no longer maintained
sqlalchemy-postgres-copy package, which we called exactly one time, I
wrote a little bit of raw SQL that uses `copy_expert` to do the same
thing, just for the single way we're using the postgres COPY FROM
command. This will allow us to package up the last postgresql dependent
version of PUDL for installation via conda, if for some reason someone
needs that... and also so we can practice that packaging process.

Progress toward #371