Alter pudl script calling to ensure passage of tests.
As with the `conda-forge` tests, this change calls `pudl_setup` first,
before attempting to run the rest of the scripts using the `--help`
flag, ensuring that there is a `.pudl.yml` file present... so the other
scripts can run.  But really we need to ensure that they run with or
without that file present.

Temporary workaround for Issue #406