Fix scripts to run w/o .pudl.yml, create user environment.yml
Fixed try/except logic in setup module to catch FileNotFoundError
correctly, and return None for pudl_in and pudl_out if there's no
default file supplied by the user.

Closes #406

Also created a separate "user" environment.yml file that is stored in
`package_data` and deployed by the `pudl_setup` script, which installed
catalystcoop.pudl using conda... or at least, it will, when the package
is available on conda-forge.  Will still need to figure out what the
right way is to bootstrap the user into that conda environment, but
that's a documentation issue.  Will also obviously need to test the
environment.yml file deployed in a clean environment.

Progress toward #383