Update instructions and example settings files to work
Now that we finally have all the packing parts in place, I was able to
walk through the legacy postgres setup instructions and see if they all
work... and of course they didn't quite.  So I updated the instructions
and some of the settings files to reflect the current reality and they
should hopefully work for new users now, and get folks to having a
year's worth of data available locally.