Update dependencies and copy edit README / config files.
We now depend on datapackage 1.9.0 so that has been updated in the
environment file and setup.py.  Added a clarification to the README that
the current getting started instructions shouldn't (quite) be expected
to work right now. Added additional comments and meaningful descriptions
to the etl_example.yml settings file. Removed obsolete `ferc_testing`
attribute.