Update documentation for v0.2.0
Went through most of the documentation and tried to make sure it is
correct as of the current codebase, for release alongside the first data
package based version of PUDL.

Also moved README.rst and LICENSE.txt into the main directory of the
repository, since some of the other repository and release tools are not
able to find them in the docs directory.  Their contents are being
included where appropriate in `index.rst` and `LICENSE.rst` in the docs.

Progress toward #401