Standardizing plant and utility name and ID column names
Attempt to standardize plant and utility name and ID field names across
all tables and datasets, using uniform _pudl, _eia, and _ferc1 suffixes.
Also remove lingering _ferc suffixes on other variables and functions,
replacing them with _ferc1 for uniformity.

In the process, we uncovered the possibility of mismatches between the
columns which exist in the transformed dataframes, and the columns
specified by the JSON metadata, added a rigorous check, and found
several actual mismatches, which have not been worked through.
Currently ETL is still broken because of these mismatches, and all of
the post-ETL functionality is b0rked because the plant & utility ID
column names can't be debugged there until the ETL process is working
again. Blargh.

Progress toward #447