Update validation notebooks to work w/ PudlTabl object
Somehow the PudlTabl object came to require that the pudl_engine be
specified by keyword only, which I think maybe was not intentional? This
broke some of the PudlTable object initalization in the existing data
validation notebooks.

Deleted some functions defined in the FRC notebook, which were
integrated into the validate module.

Updated the tox environment for notebook validation to run them all at
once, so that it doesn't fail out if one set of notebooks has an error
without running all the rest.

Fixed the gens_eia860 notebook to work when weights are not supplied.