Finishing touches on bulk data type conversion
Now there is a framework in place to convert data types of a dataframe 
or of a dictionary of dataframes. Right now this is being applied only 
to EIA 860, mostly during the transform step. This could later be 
expanded upon to incorporate all of the data sources and happen at 
various stages. Working on issue #403.