Add data validation tests for FERC Steam Plants
In addition to adding a bunch of data specific test cases for the FERC
Form 1 Steam Plants table, parametrized many of the vs_bounds tests in
the other validation modules for ease of reading.

Added capacity_factor, opex_nonfuel_per_mwh, and opex_fuel_per_mwh to
the steam plants output table.

Removed the PUDL Tables metadata requirement in the FERC output
functions -- it wasn't being used, and the kind of select statements
that we are doing in the EIA output tables are also very simple, and
don't benefit much from using SQLAlchemy (and also the metadata object
can be obtained locally by reflecting the database engine, so it
doesn't need to be passed around all over the place).

Progress toward #277