Fix excess Hg content reported in 2012 frc_eia923
As described in #390 there's a host of bad coal mercury content data
reported almost exclusively in 2012 in the EIA 923 Fuel Receipts and
Costs table. This adds a step to the transform process that replaces any
value larger than 7.0ppm (the smallest known obviously bad value in the
data to 2018) with NA.

Also added the instructions for filling out EIA 860/923 to the docs,
just to make sure we have them on hand for reference.

Closes #390