Break-out all energy sources for heat content validation
While I was in the validation cases fixing the coal mercury content, I
went ahead and broke out the other types of gas that were the issue in
issue #391 along with all of the other various fuel types, since the EIA
923 instructions had concrete upper/lower bounds on their allowed heat
contents.

See: https://www.eia.gov/survey/form/eia_923/instructions.pdf

The number of data validation cases has gotten pretty large at this
point and we probably want to think of a more compact/lower overhead
way to encode them, so we can easily keep adding them over time without
it being a drag -- and we also need to make the re-creation of them in the
validation / debugging notebooks easier. Doing it piecemeal one at a
time gets pretty annoying... annoying enough that I didn't do it here
when updating the validation cases that are being used in the tests.
I've added some notes to issue #400 about this.

Closes #391