Update EIA923 xlsx maps to deal with energy storage
EIA added a new "energy storage" (!) tab to the EIA 923 that looks a lot
like the generation_fuel tab.  Added it (and also... all the various
stocks_oil, stocks_coal, etc) tabs to the tab maps and updated a little
bit of code so that only the pages which work are pulled in by the ETL
process.