Fixing composite foriegn keys in metadata
We had several of the foriegn keys definied incorrectly and it was 
causing errors with the Postgres export. An example of the broken FKs 
was the plant_id_eia/generator_id FK. The plant_id_eia/generator_id is a 
composite PK in the generators table and we previously had the 
generator_id pointing to the generators table and the plant_id_eia 
pointing to the plants table. That's just wrongo, so I squished them 
together to point to the generators table. There were a few other 
instances of this as well.