Standardize abbreviation for "data package"
Did a bunch of search-and-replace to standardize on "datapkg" as the
shorthand for "data package" since it was being abbreviated in a bunch
of different ways in different places.

Users will need to re-name their workspace `datapackage` directories to
`datapkg` as a result.

Also changed many instances of the plural of data packages, which were
referring to data package bundles to `datapkg_bundle`.  The Travis tests
are passing locally so hopefully the necessary changes have been made
everywhere.

Also added a `version` field for the datapackage version in connection
with #425 (the data release script).

Closes #456