Add optional datapackage bundle DOI to ETL settings file
Added the option of adding a DOI to the ETL settings file, which would apply to
the entire bundle of data packages, and populate a datapkg_bundle_doi field
within each of the component data packages (alongside the datapkg_bundle_uuid
that's automatically generated to identify the ETL run).

Added logic to validate both the UUID and DOI values that are passed in, and
got rid of the default datapkg_bundle_uuid value (which would have given every
package a different bundle UUID if it were not populated).  The
datapkg_bundle_uuid and datapkg_bundle_doi values are optional. Also made the
data package "version" attribute optional -- since it really only makes sense
to include when we're trying to do a data release.

Progress toward #425