Deduplicate ETL test fixtures
Condensed duplicate ETL steps that were laid out in the pytest
fixtures so that the way to get a DB connection, and to re-generate the
SQLite DB for testing, is to use the `pudl_engine` fixture. The
`datapkg_bundle` fixture will make the data packages if need be, and
then `pudl_engine` flattens then and loads the results into SQLite.

Also added a little bit of logic to pass the `datapkg_bundle_doi`
through to the `generate_datapkg` so that it shows up in the outputs
when it is offered.

Closes #461