Implement row unpacking for Plant in Service table
Used the new row unpacking infrastructure to convert rows into columns
corresponding to the various FERC Accounts. This allows all of the rows
to be mapped to the correct fields, even as which row number they are
associated with changes over the years.  The implemented row maps cover
all available years of data (1994-2018).

The six original columns reported in this table represented the starting
balance, additions, retirements, adjustments, transfers and ending
balance for each FERC account. These "columns" have been converted into
an index column, with entries named for each of those types of amounts
(the column is called amount_type).

The FERC Account numbers are used in the column names (acctXXX) and the
various sub-categories of accounts are used as a prefix (e.g.
transmission_acctXXX_...) followed by a 1-3 word descriptive name, taken
from the row literal describing the original entry on the form.

The conversion of rows into columns meant that we had to stop using the
same format for the record_id that identifies where in the original FERC
data an entry in the PUDL DB came from.

A small number of utilities may have used an old version of the FERC
Form 1 in 2004-2005, meaning that some of the row numbers are incorrect
(when compared to other forms filed in those years). See #471 for
details.

Closes #472