Integrate 2018 FERC 1 and EIA 860/923 data
This commit brings in the ID mapping for EIA 860/923 and FERC 1 in 2018,
and also the earlier years of FERC 1 -- from 1994 through 2003.

A little bit of additional data cleaning was required to deal with the
earlier data, but overall it seems to have pretty much just worked once
the mapping was done.

Closes #344
Closes #286
Closes #428