Add rolling averages for null fuel cost in outputs
Working on Issue # 160. For our work with RMI, we wanted to fill in some 
of the missing monthly fuel cost data. Turns out, this does not do a 
whole lot for adding a lot of records. Becuase of proprietary reporting, 
a good chunk of the fuel cost records in the frc table are just 100% 
null records.

Nonetheless, now when initializing the PudlTabl object, you can set 
`rolling=True` to get a rolling average for the fuel cost. This is used 
heavily in the MCOE calcs, as thus produces more results.