Add utility_id_eia into generators_eia860
We were previously associating the utility_id_eia with a plant in the 
harvesting/normalization process. Since I learned the ownership table 
does not have all of the generators, I wanted an easy to grab utilty 
associated with all of the generators. Now utility_id_eia is in the 
normalized generators and plants tables.