Generatic(ish) excel extractor for 861.
I've added a new 861 module that we could eventually transition into a 
fully generic excel extractor. I've added a eia861 subdirectory in 
xlsx_maps directory. Hopefully this type of structure of storing data 
source metadata will be able to be used for other datasets in the 
future.

For now, the EIA 861 sales table is the only one that is working. It is 
working from 1990 through 2018.