Update tests to account for aggregated/raw outputs
Some of the data validation tests can only run on aggregated dataframes
(e.g. the MCOE), and some of them can only run on the raw dataframes.
However, the parameterization of the tests can't distinguish between the
two, because the output objects that are getting passed around and
re-used aren't instantiated until after the setup process is done.

To get around this issue I've added conditional `pytest.skip()` calls
within the tests that check the value of the `PudlTabl.freq` attribute.