Allow selective loading of datapackages into SQLite
The EPA CEMS data is too large to load into SQLite en masse, and much
easier to work with as Apache Parquet files that are read in selectively
alongside the database for use in pandas or dask dataframes. However,
the previous version of the datapackage flattening code and the script
for loading data packages into SQLite would only load *all* of the
datapackages within an ETL bundle into the database. This commit changes
the flattening code to allow the user to select some subset of the
datapackages which are part of a bundle for merging and loading into
SQLite. Because the flattening code was previously consuming a whole
datapackage bundle, this required some edits in pretty much all of the
functions in that module.

In addition the datapkg_to_sqlite script now allows the user to *only*
merge, and not load into SQLite, in case folks want to create a big
datapackage for loading into another storage medium like BigQuery.

Closes #493