Update README for 0.3.0 & switch to 2018 data for CI
Since it's all working smoothly now, and should be used for the example /
tutorial (which means it's important that we know if it *stops* working) I
switched the Travis CI test data over to 2018.

Removed a directory of settings files that were erroneously checked into
the top level repo directory in November.

Updated example YAML settings files to work with current software, including
renaming the pkg_bundle variables to datapkg_bundle, and also switched to using
2018 data, and populating both the EIA only and EIA-EPACEMS data packages, so
that users working through the example ETL process can do the EPA CEMS Parquet
conversion, and have a DB with no CEMS in it to practice with.

Updated the top level CLI module to work with current function settings (i.e.
no debug setting expected / required / allowed)

Updated the README (esp. the quickstart guide) to reflect the current usage of
the command line scripts, to refer to 2018 data, etc.