Fix column mapping and data typing errors
There were several errors in the column maps for EIA 860 which became
apparent when I attempted to debug issues with harvested value
consistency during ETL for the 2009-2010 EIA 860 data. Some of the 2018
columns for the proposed generators were missed, and there were a couple
of typos. This resulted in extra columns floating around, which were
only noticed when I tried to do exhaustive (rather than selective) type
assignment to the columns in data frames during the transform step, and
found that there were still some of the original, un-renamed column
titles floating around.

Added a test case that will ensure no ownership fractions < 0.0 or >
1.0.

Still can't call the 2009-2010 EIA 860 data integration (Issue #484)
done because there are ~2000 additional plants which appear only in
those older years (many of which are 100-1000 MW in scale) which need to
be mapped to their FERC equivalents and/or assigned PUDL IDs.

Progress toward Issue #484 (Adding 2009-2010 EIA 860 data)