Refinements to the data validation process & output objects
Increased the allowed proportion of generators with partial ownership records
to 0.5% due to more partial records existing when we don't drop NA as
aggressively.

Change the method of checking whether a table has the expected number of rows
to use an "expected" number, and a +/- margin, 0.05 (aka 5%) by default.

Updated all of the expected numbers of rows/records for the EIA 860/923, MCOE
and FERC 1 tables that we're checking.

Changed output functions for EIA 860 and 923 to:
  * Only dropana() on the subset of columns which are primary keys for the
    table being output.
  * Use Int64 nullable integer types for all the plant_id_* and utility_id_*
    columns

Closes #530