Update to heat_rate_by_gen
I was playing around with the heat_rate_by_gen function and noticed that you define gens_simple and merge bga_gens with it, but then you do not use gens_simple again in the function. I am wondering if you meant to call this merge bga_gens instead, since you do use bga_gens in line 110 when defining your hr_by_gen dataframe. (Not sure if this is the proper way to initiate a pull request so please let me know if there is a different protocol)