Test with Python 3.8 and remove deprecated kwargs
Removed the apparently unneeded "box" kwarg from the call to
`pd.to_timdelta()` that was happening in fixing up the EPA CEMS
timestamps. Only difference is the return type, and it's being added to
a `pd.datetime` so I'm going to assume that Pandas is doing the Right
Thing™ here.

The Packaging issues that were causing a `pip` based install of PUDL to
fail on Python 3.8 seem to have... fixed themselves? Appears to be
working now without any additional intervention. Maybe it was an issue
in some other dependency that got fixed.

Closes #549
Closes #537