Initial attempt to make PUDL work with Pandas 1.0
The "box" issue and the missing indices in .loc[] for the FERC Form 1
plant ID assignment were easy to fix. A slightly more annoying case came
up where the string "None" appeared instead of the None type in the
Timezone strings used with CEMS.

However, it looks like there are now a whole lot of "nan" and "None" and
such strings floating around, breaking a bunch of our data cleaning
logic -- the ETL runs now, but fails on data validation because these
fake "NA" values are of course not in the enumerated lists of acceptable
values in the datapackage.json metadata (thank goodness for ENUMs!).

So my guess is it's more worthwhile to just dive into tackling #252 than
it is to try and chase down all of these issues one at a time to patch
them temporarily.

Closes #549
Closes #222