fuel cost output to ref 860 generators.
In investigating null fuel cost calcs, I narrowed in on the fuel cost 
output func being a source of the loss of fuel cost values. This 
function used to use the generation table from 923, probably as a 
vestige of having developed a lot of these output functions before we 
really integrated 860.