Extractor/metadata for EIA 861.
All of the "pages" of 861 should now be extractable with 
`pudl.extract.eia861.Extractor().extract` for 2010 through 2018. In 
order to deal with the fact that the colum maps/renaming is different 
for 861 (which uses numeric locations instead of strings) I pulled the 
standard renaming into `process_raw` and overwrote that in the 861 
version. I thought about creating a generic rename function and 
overwriting that but seemed like overkill - although now all of the EIA 
extractors have to rename their columns so maybe this method would be 
better. I'm very open to thoughts 
on this! I also pulled up the testing argument into 
`GenericExtractor.extract` 
so it was easier to turn that knob. I also added a way to skip years of 
pages where there is no filename. Right now this is just when the 
file_map.csv has a -1 for that year. I don't love this - I'd love 
feedback here. Working on #590.

Most of the work here was metadata munging....

Also, Close #470. All of the eia metadata is transposed and the 
GenericExtractor knows how to read the transposed metadata.

WIP: The `pudl.extract.excel_filename` was having a hard time finding a 
few of the files pre-2009. I need to figure that out.

WIP: `pudl.extract.load_excel_file` was caching the first page's file 
and returning that. Not sure what the right solution is.