Initial EIA 861 and refined FERC 714 transform functions.
The FERC 714 demand time series is now complete -- meaning there are no missing
timesteps, and no duplicate timesteps. Now we can move on to using the
anomolous value identification routines developed by @truggles and parallelized
by @gschivley to identify other values which need to be removed and imputed. See
the pudl.transform.ferc714._complete_demand_timeseries() function.

Closes #588

Added pudl.analysis.demand_mapping and pudl.transform.eia861 modules to the
automatic top level __init__.py imports.

Started adding transform functions for EIA 861, beginning with the
service_territory_eia861 table, which primarily entails adding the county FIPS
codes at the moment. As of yet there are no other transform functions.

Added some helper functions for cleaning up the county names as reported to
EIA and associating the county FIPS codes with them by name and state. These
will be useful for both the EIA 861 and the EIA 860. This also resulted in a
new dependency, on the addfips package, which appears semi-abandoned. It's a
single module package though, and we might want to re-write it to do some
fuzzy matching based on names in the future. It is also not available via
conda-forge, so this is not an ideal situation.

Progress toward #596