Reduce complexity of FERC 714 transform process.
Early on I was overly optimistic about the information content of the
FERC 714 data, and built some of the time series clean up assuming that
the offset codes (EDT, PST, etc.) were indicative of the UTC offset
associated with the data reported. It turned out that wasn't true --
FERC's instructions say all the data is to be reported in Standard Time,
and so far as I can tell that's generally what utilities do.

Similarly, of all the 25th hours that are reported, only 110 of them
were both non-zero and not daily totals, and given that the whole day is
supposed to be reported as a single database record, associated with a
single UTC offset, which is always supposed to be Standard Time anyway,
the 25th hours should, I think, never actually contain meaningful
"spring forward" time change related data.

After discovering all that, I was eventually dropping all the 25th hours
anyway, so I removed the infrastructure that was originally there to do
that exploration and reporting.

Also simplified the process of flipping the sign on a couple of
utility-years worth of demand reported as negative numbers, and the
brute force setting of some year-dependent UTC offset codes, since there
are only a few of them.

Closes #588