Further generalization of the demand mapping module
A little bit of docstring and argument naming cleanup in some of the
demand mapping functions.  Still more to be done.

Also added a couple of packages to the pudl-dev environment: lxml is
required to use the pandas.read_html() function, and pysal is a spatial
analysis library used to make the maps in the demand_mapping notebook