WIP: make gens 860 tbl the backbone of pudl_out.mcoe
We need to make a decision about how to merge annual records with 
monthly data tables in which the entities don't show up in the data 
tables. Should we make a complete monthly time series w/ nulls for every 
month? Probably, but where and how should we implement that.

Also converting the mcoe filters to null records instead of drop them.