Refine draft EIA 861 & FERC 714 transform processes
This is a WIP commit that starts integrating the EIA 861 BA and Sales
tables into Phase 2 of the Transform step (entity harvesting), but there
are some changes we need to make to that process in order for it to
actually work with data that's structured like the EIA 861 is.

A couple of functions (for pulling in US Census DP1 and HIFLD Planning
Area geometries) are also added into the demand_mapping module (one is
new, the other is moved from the ferc714 transform module where it
didn't really belong.

There were a bunch of minor EIA 861 metadata changes that resulted from
actually getting more familiar with the data.

Added the _eia suffix to balancing_authority_{id, code, name} columns
since this type of entity also shows up in the FERC 714 data, and in
multiple EIA datasets.